+++ packages/components/src/Form/Fields/FieldTime/FieldTime.tsx
@@ -23,23 +23,40 @@
  SOFTWARE.
 
  */
-import React, { FC } from 'react'
-import { render } from 'react-dom'
-import { ComponentsProvider } from '@looker/components'
-import { ButtonSetDemo } from './Actions/ButtonSetDemo'
 
-const App: FC = () => {
-  return (
-    <ComponentsProvider ie11Support>
-      <AccordionDemo />
-    </ComponentsProvider>
-  )
-}
+import React, { forwardRef, Ref } from 'react'
+import styled from 'styled-components'
+import { useID } from '../../../utils'
+import { useFormContext } from '../../Form'
+import { InputTime, InputTimeProps } from '../../Inputs/InputTime'
+import { Field, FieldProps, omitFieldProps, pickFieldProps } from '../Field'
 
-/*
-  This is the binding site for the playground. If you want to edit the
-  primary application, do your work in App.tsx instead.
- */
-document.addEventListener('DOMContentLoaded', () => {
-  render(<App />, document.getElementById('container'))
-})
+export interface FieldInputTimeProps extends FieldProps, InputTimeProps {}
+
+const FieldTimeComponent = forwardRef(
+  (props: FieldInputTimeProps, ref: Ref<HTMLInputElement>) => {
+    const validationMessage = useFormContext(props)
+    const id = useID(props.id)
+    return (
+      <Field
+        {...pickFieldProps(props)}
+        id={id}
+        validationMessage={validationMessage}
+      >
+        <InputTime
+          {...omitFieldProps(props)}
+          aria-describedby={`${id}-describedby`}
+          id={id}
+          validationType={validationMessage && validationMessage.type}
+          ref={ref}
+        />
+      </Field>
+    )
+  }
+)
+
+FieldTimeComponent.displayName = 'FieldTimeComponent'
+
+export const FieldTime = styled(FieldTimeComponent)`
+  width: 100%;
+`